+++ config/symbols.us.stnz0.txt
@@ -23,7 +23,7 @@ g_UnkRecursPrimVecOrder = 0x80181F40;
 g_UnkRecursPrim2Inds = 0x80181F60;
 g_eBlueDoorUV = 0x801826B8;
 g_eBlueDoorTiles = 0x801826D0;
-EntityBreakable = 0x801B0EEC;
+EntityBreakableNZ0 = 0x801B0EEC;
 EntityRedEyeBust = 0x801B11C0;
 EntityPurpleBrickScrollingBackground = 0x801B12E8;
 EntityLeftSecretRoomWall = 0x801B14C4;
@@ -79,3 +79,4 @@ EntityBloodSkeleton = 0x801C7958;
 EntityMagicallySealedDoor = 0x801C7D68;
 EntityStageNamePopup = 0x801C8CAC;
 EntityLifeUpSpawn = 0x801CA160;
+g_eBreakableInit = 0x80180BC8;
